+++ src/render/texture.cpp
@@ -161,13 +161,13 @@ void Texture::generateMipmaps(int width, int height, const VkCommandPool & comma
 
         VkImageBlit blit = {};
         blit.srcOffsets[0] = { 0, 0, 0 };
-        blit.srcOffsets[1] = { mipWidth, mipHeight, 1 };
+        blit.srcOffsets[1] = {(int32_t) mipWidth, (int32_t) mipHeight, 1 };
         blit.srcSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
         blit.srcSubresource.mipLevel = i - 1;
         blit.srcSubresource.baseArrayLayer = 0;
         blit.srcSubresource.layerCount = 1;
         blit.dstOffsets[0] = { 0, 0, 0 };
-        blit.dstOffsets[1] = { mipWidth > 1 ? mipWidth / 2 : 1, mipHeight > 1 ? mipHeight / 2 : 1, 1 };
+        blit.dstOffsets[1] = { (int32_t) (mipWidth > 1 ? mipWidth / 2 : 1), (int32_t) (mipHeight > 1 ? mipHeight / 2 : 1), 1 };
         blit.dstSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
         blit.dstSubresource.mipLevel = i;
         blit.dstSubresource.baseArrayLayer = 0;
